address build failures

+++ rift/ads/features/stream_features/content_keyword_clicks_push.py
@@ -1,6 +1,6 @@
 from datetime import timedelta, datetime
 from tecton import StreamFeatureView, BatchTriggerType, Attribute
-from tecton.types import Int64
+from tecton.types import Int32
 from ads.entities import content_keyword
 from ads.data_sources.ad_impressions import keyword_click_source
 
@@ -25,6 +25,6 @@ content_keyword_click_counts_push = StreamFeatureView(
     timestamp_field='timestamp',
     environment='tecton-core-1.1.0',
     features=[
-        Attribute(name='clicked', dtype=Int64),
+        Attribute(name='clicked', dtype=Int32),
     ]
 )
