+++ rift/ads/features/stream_features/user_ad_impression_counts.py
@@ -1,5 +1,5 @@
 from tecton import stream_feature_view, Aggregate, AggregationLeadingEdge
-from tecton.types import Int64, Field
+from tecton.types import Int32, Field
 from ads.entities import ad, user
 from ads.data_sources.ad_impressions import ad_impressions_stream
 from datetime import datetime, timedelta
@@ -10,9 +10,9 @@ from datetime import datetime, timedelta
     entities=[user, ad],
     mode='pandas',
     features=[
-        Aggregate(input_column=Field('impression', Int64), function='count', time_window=timedelta(hours=1)),
-        Aggregate(input_column=Field('impression', Int64), function='count', time_window=timedelta(hours=24)),
-        Aggregate(input_column=Field('impression', Int64), function='count', time_window=timedelta(hours=72)),
+        Aggregate(input_column=Field('impression', Int32), function='count', time_window=timedelta(hours=1)),
+        Aggregate(input_column=Field('impression', Int32), function='count', time_window=timedelta(hours=24)),
+        Aggregate(input_column=Field('impression', Int32), function='count', time_window=timedelta(hours=72)),
     ],
     timestamp_field='timestamp',
     online=False,
