+++ rift/ads/features/stream_features/user_click_counts_push.py
@@ -1,6 +1,6 @@
 from datetime import timedelta, datetime
 from tecton import StreamFeatureView, Aggregate
-from tecton.types import Bool, Field, Int64
+from tecton.types import Bool, Field, Int32
 from ads.entities import user
 from ads.data_sources.ad_impressions import user_click_push_source
 
@@ -17,9 +17,9 @@ user_click_counts_push = StreamFeatureView(
     feature_start_time=datetime(2023, 1, 1),
     alert_email="demo-user@tecton.ai",
     features=[
-        Aggregate(input_column=Field('clicked', Int64), function='count', time_window=timedelta(hours=1)),
-        Aggregate(input_column=Field('clicked', Int64), function='count', time_window=timedelta(hours=24)),
-        Aggregate(input_column=Field('clicked', Int64), function='count', time_window=timedelta(hours=72)),
+        Aggregate(input_column=Field('clicked', Int32), function='count', time_window=timedelta(hours=1)),
+        Aggregate(input_column=Field('clicked', Int32), function='count', time_window=timedelta(hours=24)),
+        Aggregate(input_column=Field('clicked', Int32), function='count', time_window=timedelta(hours=72)),
     ],
     timestamp_field='timestamp',
     tags={'release': 'production'},
